requirements.txt
fastapi==0.115.0
praw==7.7.1
openai==1.52.0
python-dotenv==1.0.1
uvicorn==0.30.6